Refactor freetype fix 3
